helm: do not underline selection
Instead of underlining the selection use a more visible background
color (+1 instead of -1, +2 might also be an option).  This is more
visible because it is different from the background color of section
headers.

Please note that it is unclear whether `helm-selection-line' does have
any effect at all; it appears `helm-selection' always wins.