linum: use same background color as in fringe
This addresses #38 though but it does not actually do what that asked
for.  Giving the fringe the same background color as the linum thingy
(but different from that of default) would require that a function is
added to `linum-mode-hook` to change the fringe face locally in the
current buffer.

I admit the later would look better.  But it is also more confusing.
So let's not do it.