magit-item-highlight: leave :bold unset
Explicitly setting `magit-item-highlight' to nil was only necessary for
the short period during which magit highlighted the current section by
making it bold.  Since that no longer is the case the only effect of
explicitly setting the highlighted section to being not-bold is that
parts of a section loose their boldness when it is current.