adapt to magit v2.1.0
Besides the `magit', `git-commit', and `git-rebase' faces, this also
adjusts the `diff' faces.  Previously the `diff' faces did explicitly
_not_ set the background color.  But that was a mistake.  Diff
refinement faces _have to_ set the background, because they are also
used to highlight changes in whitespace and that wouldn't be visible
without a background color different from that of `default'.  The other
`diff' faces should also set the background color because otherwise the
refinements would look very ugly.  And in my opinion diffs look better
with a background color even when no refinements are involved.