Add a mechanism to override a subset of the theme's colours
If a user wants to tweak the theme by changing just a few colours, they
can now do so by defining new values for just that subset in the new
"zenburn-override-colors-alist" variable.

This improves over the previous approach of having to completely
re-specify the full list of colours, which both obscures the actual
overrides among the unchanged defaults, and risks breaking when new
colours are added to the default theme without the user updating their
definition.