magit, diff: Cosmetics whitespace changes
Make the definitions of Magit's Zenburn faces look like those of the
Solarized theme.  I now use the latter but used to use the former,
and making this look similar for both themes makes it easier for me
to fix the regressions introduced by others.

Do the same for the diff faces.