Uncomment and fix the advices that add zenburn-specific font-lock
I originally implemented these advices in #106.  They got commented
out in 27cee3d in response to #110 and similar reports in prelude's
issue tracker.

I just implemented the same thing again, only to find the commented
advices once I was done... five years later... it seem I would very
much like this.

So I had a look at the reports and it seems pretty obvious what the
issue was: just because the major-mode is `emacs-lisp-mode', that
does not mean that the buffer is visiting a file.  And the fix is
to check whether the buffer is visiting a file before trying to do
something with the `buffer-file-name'.

Besides reverting the commenting and fixing the bug, this commit
also does the following:

- Use `zenburn-default-colors-alist' because that variable has since
  been renamed.

- Add out keywords at the end because otherwise the "blue" in
  "zenburn-blue" for example would have "blue" as the background color
  instead of "zenburn-blue".

- Extend the doc-string of `zenburn-add-font-lock-keywords' to make
  users aware of a complication.