Add zenburn-theme to faces Custom group
I don't think it is customary to add theme groups to any parent group,
but that's probably just because themes are usually not byte-compiled,
so the byte-compiler didn't get a chance to inform the author about
this issue.

The `faces' group seems like an obvious parent.  I wouldn't worry
about polluting that group with lots of theme sub-groups if this
catches on because for that to happen all the libraries that define
those themes would have to be loaded and if one did that, then one
would have bigger problems than to many sub-groups.

On the other hand, if one does use the `zenburn' theme (or another
theme), then it is nice to find that in `faces'.