Revert "Modify diff colors"
This reverts commit 8856da44c58193bd6d05b8fd8cfd0c876a12fc87 from
pull-request #300.

Also see #317.

The goal of that commit was to make the diffs prettier by making them
darker, which I agree was a success.  Unfortunately the diffs not only
got prettier but also lost information.

The `diff-refine-*' faces quite obviously must be different from the
`magit-diff-*-highlight' faces and the commit being reverted violated
that in the most explicit way possible by making the latter inherit
from the former.

The refinement or "word-level" faces are used to change the parts of
a line that changed.  The highlighting faces are used to emphasize
the "selected" part of a diff, e.g. the part of the diff the user is
currently concentrating on by placing the cursor inside that part of
the diff.  Obviously if one concentrates on a part of a diff, then
one does not want that part to become *less* informative until one
focuses elsewhere.

I tried fixing this regression without going as far as to completely
reverting the faulty commit, unfortunately I was unable to do so.